Skip notebook addon in browser case
